modifications

+++ audio_coding/codecs/isac/fix/source/arith_routines_logist.c
@@ -226,7 +226,7 @@ int WebRtcIsacfix_EncLogisticMulti2(Bitstr_enc *streamData,
  * Return value             : number of bytes in the stream so far
  *                            -1 if error detected
  */
-int16_t WebRtcIsacfix_DecLogisticMulti2(int16_t *dataQ7,
+int WebRtcIsacfix_DecLogisticMulti2(int16_t *dataQ7,
                                         Bitstr_dec *streamData,
                                         const int32_t *envQ8,
                                         const int16_t lenData)
